// frontend/src/api/models/ModelsWebhookRequest.ts
// @ts-nocheck
/**
 * Mock Payment Provider API
 * API for mocking payment provider integrations with webhooks, 3DS, and admin dashboard.
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export class ModelsWebhookRequest {
    'active'?: boolean;
    'events'?: Array<string>;
    'url'?: string;

    static readonly discriminator: string | undefined = undefined;

    static readonly mapping: {[index: string]: string} | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "active",
            "baseName": "active",
            "type": "boolean",
            "format": ""
        },
        {
            "name": "events",
            "baseName": "events",
            "type": "Array<string>",
            "format": ""
        },
        {
            "name": "url",
            "baseName": "url",
            "type": "string",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return ModelsWebhookRequest.attributeTypeMap;
    }

    public constructor() {
    }
}